Factor encoding functions and add separate topic encoding
