+++ packages/codec/lib/basic/index.ts
@@ -5,3 +5,11 @@ export {
    */
   Decode
 };
+
+import * as Encode from "./encode";
+export {
+  /**
+   * @protected
+   */
+  Encode
+};
