+++ packages/codec/lib/bytes/index.ts
@@ -13,3 +13,11 @@ export {
    */
   Decode
 };
+
+import * as Encode from "./encode";
+export {
+  /**
+   * @protected
+   */
+  Encode
+};
